# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Which Docs
# during dev, should point to your local machine
host: localhost:5068
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/getTitleList:
    # binds a127 app logic to a route
    x-swagger-router-controller: title_ctrl
    post:
      tags: ['Title Module']
      description: Get Title List
      summary: Get Title List
      # used as the method name of the controller
      operationId: getTitleList
      parameters:
        - name: body
          in: body
          description: Get Title List
          schema:
           $ref: '#/definitions/getTitleList'
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/getFrontdeskTitles:
    # binds a127 app logic to a route
    x-swagger-router-controller: frontdesktitle_ctrl
    post:
      tags: ['Title Module']
      description: Get Title List
      summary: Get Title List
      # used as the method name of the controller
      operationId: getFrontdeskTitles
      parameters:
        - name: body
          in: body
          description: Get Title List
          schema:
           $ref: '#/definitions/getFrontdeskTitles'
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"            
...
  /api/addFrontdeskTitle:
    # binds a127 app logic to a route
    x-swagger-router-controller: frontdesktitle_ctrl
    post:
      tags: ['Frontdesk Title Module']
      description: save Title
      summary: Add new Title
      # used as the method name of the controller
      operationId: addFrontdeskTitle
      parameters:
        - name: body
          in: body
          description: sets Frontdesk title name
          schema:
           $ref: '#/definitions/addFrontdeskTitle'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"          
...
  /api/getFrontdeskTitle/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: frontdesktitle_ctrl
    get:
      tags: ['Frontdesk Title Module']
      description: Get Title By Id
      summary: Get Title By Id
      # used as the method name of the controller
      operationId: getFrontdeskTitleById
      parameters:
        - name: id
          type: string
          in: path
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/deleteFrontdeskTitle:
    # Doctors APIs
    x-swagger-router-controller: frontdesktitle_ctrl
    post:
      tags: ['Frontdesk Title Module']
      description: delete Title
      summary: delete Title
      # used as the method name of the controller
      operationId: deleteFrontdeskTitle
      parameters:
        - name: name
          in: body
          description: delete Title
          schema:
            $ref: '#/definitions/deleteFrontdeskTitle'
          required: true
          # type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"                     
...
definitions:
...
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      message:
        type: string
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  getTitleList:
    type: object
    properties:
      page:
        type: string
      count:
        type: string
  getFrontdeskTitles:
    type: object
    properties:
      page:
        type: string
      count:
        type: string     
  addFrontdeskTitle: 
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      desc:
        type: string
  deleteFrontdeskTitle:
    type: object
    properties:
      id:
        type: string
...